Ajout condition lors de la modification du profil

+++ scripts/install.sh
@@ -176,15 +176,17 @@ if [ $? == 0 ] ; then
 	cp /opt/borne/scripts/sessionStart.desktop /etc/xdg/autostart/sessionStart.desktop
 
 	#Copie des profils
-	cp /opt/borne/share/skel_admin.tar.gz /home/
-	cp /opt/borne/share/skel_gest.tar.gz /home/
-	cd /home/
-	rm -rf gestionnaire/
-	tar -xvzf skel_gest.tar.gz
-	mv skel_gest/ gestionnaire/
-	rm -rf administrateur/
-	tar -xvzf skel_admin.tar.gz
-	mv skel_admin/ administrateur/
+	if [ $1 != "maj" ] ; then
+		cp /opt/borne/share/skel_admin.tar.gz /home/
+		cp /opt/borne/share/skel_gest.tar.gz /home/
+		cd /home/
+		rm -rf gestionnaire/
+		tar -xvzf skel_gest.tar.gz
+		mv skel_gest/ gestionnaire/
+		rm -rf administrateur/
+		tar -xvzf skel_admin.tar.gz
+		mv skel_admin/ administrateur/
+	fi
 
 	echo "Fin de l'installation"
 fi
